+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Support Souffle interpreter for development.
 * Add support for loading existing GTIRB files.
 
 # 1.4.0